// src/app/models/entities.ts
export interface Field { 
    ColumnName?: string;
    IsKey?: boolean;
    IsNullable?: boolean;
    MaxLength?: number;
    ReadOnly?: boolean;
    RegularExpression?:string;
}
// ...
export interface SearchSortRequest { 
    field?: string;
    dir?: string;
}

export interface SelectItem { 
    value?: any;
    label?: string;
}